requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-telegram-bot==20.7
aiohttp==3.9.3
